Move MatchesOperator to pylot.utils
